requirements.txt
pandas>=1.5.0
pyyaml>=6.0.0
pydantic>=2.0.0 